LayoutTests/js/Promise-exception-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInThenCallback === thirdPromise is true
PASS result is "foobar"
PASS successfullyParsed is true

TEST COMPLETE
